Add files via upload
In this project I was calculating the loads on an airplane wing using data provided for the students. The assumption was made that all the forces were acting on a wing spar who dimensions were also provided with an excel file. I started by creating bendingStress function,I then calculated the curvature of the wing which I also turned into another function. This data was then store in a structure array called wing data where this data was separated into two fields. The force fields and moment fields.